+++ src/components/layout/header.tsx
@@ -6,17 +6,17 @@ import { Separator } from "@/components/ui/separator";
 
 export function Header() {
   return (
-    <header className="z-10 flex lg:h-20 items-center border-b border-border bg-card text-sm">
+    <header className="z-10 flex items-center border-b border-border bg-card text-sm lg:h-20">
       <div className="mx-auto w-full max-w-7xl">
-        <div className="flex flex-col lg:flex-row items-center justify-between gap-y-4 p-4 lg:py-0">
+        <div className="flex flex-col items-center justify-between gap-y-4 p-4 lg:flex-row lg:py-0">
           <div className="flex items-center">
             <Logo />
 
-            <Separator orientation="vertical" className="mx-6" />
+            <Separator orientation="vertical" className="mx-6 hidden sm:block" />
 
             <HeaderNav />
           </div>
-          <div className="flex items-center gap-4">
+          <div className="grid items-center gap-4 sm:grid-cols-[1fr_auto]">
             <DownloadButton />
             <HeaderUser />
           </div>